+++ changelog.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
-- Minimum time delay between auto-trim (#1)
+- Setting to control minimum time delay between auto-trim (#1)
 
 ### Changed